refactor: RoomescapeController를 HomeController로 이름 변경 및 controller 패키지로 이동
